Move extract utils into two separate helpers
