+++ src/helpers/index.js
@@ -1,4 +1,4 @@
 export filterNotification from "./filter-notification";
 export requestExtract from "./request-extract";
-// export setUserSegments from "./set-user-segments";
+export handleExtract from "./handle-extract";
 export updateSettings from "./update-settings";
